# docs/api/server.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8099",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/user/register": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespString"
                }
              }
            }
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespUserLoginResponse"
                }
              }
            }
          }
        }
      }
    },
    "/speech/recognize": {
      "post": {
        "tags": [
          "speech-recognition-controller"
        ],
        "operationId": "recognizeAudio",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "required": [
                  "audioFile"
                ],
                "type": "object",
                "properties": {
                  "audioFile": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespString"
                }
              }
            }
          }
        }
      }
    },
    "/ai/travel-plan": {
      "post": {
        "tags": [
          "ai-recognition-controller"
        ],
        "operationId": "getTravelPlan",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespString"
                }
              }
            }
          }
        }
      }
    },
    "/ai/recognize-speech": {
      "post": {
        "tags": [
          "ai-recognition-controller"
        ],
        "operationId": "recognizeSpeech",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespString"
                }
              }
            }
          }
        }
      }
    },
    "/ai/chat": {
      "post": {
        "tags": [
          "ai-recognition-controller"
        ],
        "operationId": "doChat",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespString"
                }
              }
            }
          }
        }
      }
    },
    "/ai/analyze-budget": {
      "post": {
        "tags": [
          "ai-recognition-controller"
        ],
        "operationId": "analyzeBudget",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespString"
                }
              }
            }
          }
        }
      }
    },
    "/user/profile": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "operationId": "getProfile",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RespUserProfileResponse"
                }
              }
            }
          }
        }
      }
    },
    "/ai/chat/sse-emitter": {
      "get": {
        "tags": [
          "ai-recognition-controller"
        ],
        "operationId": "doChatWithSseEmitter",
        "parameters": [
          {
            "name": "request",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ChatRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SseEmitter"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "UserRegisterRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        }
      },
      "RespString": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "string"
          }
        }
      },
      "UserLoginRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "RespUserLoginResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/UserLoginResponse"
          }
        }
      },
      "UserLoginResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ChatRequest": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "chatId": {
            "type": "string"
          }
        }
      },
      "RespUserProfileResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/UserProfileResponse"
          }
        }
      },
      "UserProfileResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        }
      },
      "SseEmitter": {
        "type": "object",
        "properties": {
          "timeout": {
            "type": "integer",
            "format": "int64"
          }
        }
      }
    }
  },
  "x-openapi": {
    "x-setting": {
      "customCode": 200,
      "language": "zh-CN",
      "enableSwaggerModels": true,
      "swaggerModelName": "Swagger Models",
      "enableReloadCacheParameter": false,
      "enableAfterScript": true,
      "enableDocumentManage": true,
      "enableVersion": false,
      "enableRequestCache": true,
      "enableFilterMultipartApis": false,
      "enableFilterMultipartApiMethodType": "POST",
      "enableHost": false,
      "enableHostText": "",
      "enableDynamicParameter": false,
      "enableDebug": true,
      "enableFooter": true,
      "enableFooterCustom": false,
      "enableSearch": true,
      "enableOpenApi": true,
      "enableHomeCustom": false,
      "enableGroup": true,
      "enableResponseCode": true
    },
    "x-markdownFiles": []
  }
}